requirements.txt
# etteytlus — Runtime Dependencies

# Audio capture
sounddevice>=0.4.6
numpy>=1.24
scipy>=1.11

# Whisper transcription (CTranslate2 backend)
faster-whisper>=1.0.0

# LLM API clients
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# Web display
flask>=3.0

# Config
python-dotenv>=1.0
